Move incorrectly placed tests
